requirements.txt
Django==4.2.*
django-admin-interface
django-allauth[socialaccount]
django-crispy-forms
django-debug-toolbar
django-htmx
django-impersonate
django-picklefield
django-ratelimit
django-redis
django-sendfile2
django-toggle-switch-widget
django-eventstream
django-mathfilters
django-constance[redis]
pypeg2
dj-database-url
crispy-bulma
python-dateutil==2.6.1
gunicorn
coverage
psycopg2==2.9.*
redis>=4.0.0
huey>=2.4.0
django-anymail
python3-openid